+++ app/Http/Controllers/IncentiveController.php
@@ -10,6 +10,7 @@ use App\Models\Employee;
 use App\Models\Incentive;
 use App\Models\PayrollPeriod;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class IncentiveController extends Controller
@@ -19,6 +20,8 @@ class IncentiveController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Incentive::class);
+
         $incentives = Incentive::with(['payroll_period', 'employees','employees.user', 'employees.position', 'employees.branch' ])->get();
         return Inertia::render('incentives/index', compact('incentives'));
     }
@@ -28,6 +31,7 @@ class IncentiveController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Incentive::class);
         $payroll_periods = PayrollPeriod::all();
         $employees = Employee::with('user')->where('employee_status', 'active')->get();
         return Inertia::render('incentives/create', compact('payroll_periods','employees' ));
@@ -38,6 +42,7 @@ class IncentiveController extends Controller
      */
     public function store(StoreIncentiveRequest $request, CreateNewIncentive $incentive )
     {
+        Gate::authorize('create', Incentive::class);
         $incentive->create($request->validated());
         DB::commit();
         return redirect()->route('incentives.index');
@@ -56,7 +61,7 @@ class IncentiveController extends Controller
      */
 public function edit(Incentive $incentive)
 {
-    
+    Gate::authorize('update', $incentive);   
     $incentive->load('payroll_period', 'employees');
     $employees = Employee::with('user')->where('employee_status', 'active')->get();
     
@@ -75,6 +80,7 @@ public function edit(Incentive $incentive)
      */
     public function update(UpdateIncentiveRequest $request,Incentive $incentive,  UpdateIncentive $updateincentive)
     {
+        Gate::authorize('update', $incentive);
         $updateincentive->update($request->validated(), $incentive);
         DB::commit();
         return redirect()->route('incentives.index');
@@ -85,6 +91,8 @@ public function edit(Incentive $incentive)
      */
     public function destroy(Incentive $incentive)
     {
+        Gate::authorize('delete', $incentive);
+
         $incentive->delete();
         DB::commit();
         return redirect()->route('incentives.index');
